# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "AssignRequest": {
        "additionalProperties": false,
        "properties": {
          "actor": {
            "default": "analyst",
            "title": "Actor",
            "type": "string"
          },
          "assignee": {
            "title": "Assignee",
            "type": "string"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          }
        },
        "required": [
          "assignee"
        ],
        "title": "AssignRequest",
        "type": "object"
      },
      "CommentRequest": {
        "additionalProperties": false,
        "properties": {
          "actor": {
            "default": "analyst",
            "title": "Actor",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text"
        ],
        "title": "CommentRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Targeted, not generic: prioritized plan-conversion validation driven by the platform pair's Migration Fingerprint. POC layer \u2014 no authentication (product adds Cognito role scopes, spec \u00a718.4); runs execute synchronously in-process (product enqueues to SQS).",
    "title": "MAPTIVA Migration Fingerprint API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/findings/{finding_id}": {
      "get": {
        "operationId": "get_finding_findings__finding_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "finding_id",
            "required": true,
            "schema": {
              "title": "Finding Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Finding"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Finding",
        "tags": [
          "findings"
        ]
      }
    },
    "/findings/{finding_id}/assign": {
      "post": {
        "operationId": "post_assign_findings__finding_id__assign_post",
        "parameters": [
          {
            "in": "path",
            "name": "finding_id",
            "required": true,
            "schema": {
              "title": "Finding Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowEvent"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Assign",
        "tags": [
          "workflow"
        ]
      }
    },
    "/findings/{finding_id}/close": {
      "post": {
        "operationId": "post_close_findings__finding_id__close_post",
        "parameters": [
          {
            "in": "path",
            "name": "finding_id",
            "required": true,
            "schema": {
              "title": "Finding Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowEvent"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Close",
        "tags": [
          "workflow"
        ]
      }
    },
    "/findings/{finding_id}/comment": {
      "post": {
        "operationId": "post_comment_findings__finding_id__comment_post",
        "parameters": [
          {
            "in": "path",
            "name": "finding_id",
            "required": true,
            "schema": {
              "title": "Finding Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowEvent"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Comment",
        "tags": [
          "workflow"
        ]
      }
    },
    "/findings/{finding_id}/history": {
      "get": {
        "operationId": "get_finding_history_findings__finding_id__history_get",
        "parameters": [
          {
            "in": "path",
            "name": "finding_id",
            "required": true,
            "schema": {
              "title": "Finding Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WorkflowEvent"
                  },
                  "title": "Response Get Finding History Findings  Finding Id  History Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Finding History",
        "tags": [
          "workflow"
        ]
      }
    },
    "/findings/{finding_id}/resolve": {
      "post": {
        "operationId": "post_resolve_findings__finding_id__resolve_post",
        "parameters": [
          {
            "in": "path",
            "name": "finding_id",
            "required": true,
            "schema": {
              "title": "Finding Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowEvent"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Resolve",
        "tags": [
          "workflow"
        ]
      }
    },
    "/findings/{finding_id}/review": {
      "post": {
        "operationId": "post_review_findings__finding_id__review_post",
        "parameters": [
          {
            "in": "path",
            "name": "finding_id",
            "required": true,
            "schema": {
              "title": "Finding Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Review",
        "tags": [
          "learning"
        ]
      }
    },
    "/fingerprints/{pair_id}": {
      "get": {
        "operationId": "get_fingerprint_fingerprints__pair_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "pair_id",
            "required": true,
            "schema": {
              "title": "Pair Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Fingerprint"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Fingerprint",
        "tags": [
          "fingerprints"
        ]
      }
    },
    "/fingerprints/{pair_id}/learning-history": {
      "get": {
        "operationId": "get_learning_history_fingerprints__pair_id__learning_history_get",
        "parameters": [
          {
            "in": "path",
            "name": "pair_id",
            "required": true,
            "schema": {
              "title": "Pair Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "fm",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Fm"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/LearningEvent"
                  },
                  "title": "Response Get Learning History Fingerprints  Pair Id  Learning History Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Learning History",
        "tags": [
          "learning"
        ]
      }
    },
    "/fingerprints/{pair_id}/publish": {
      "post": {
        "operationId": "post_publish_fingerprints__pair_id__publish_post",
        "parameters": [
          {
            "in": "path",
            "name": "pair_id",
            "required": true,
            "schema": {
              "title": "Pair Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublishAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Publish",
        "tags": [
          "learning"
        ]
      }
    },
    "/fingerprints/{pair_id}/suite": {
      "get": {
        "operationId": "get_suite_fingerprints__pair_id__suite_get",
        "parameters": [
          {
            "in": "path",
            "name": "pair_id",
            "required": true,
            "schema": {
              "title": "Pair Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "version",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Version"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuiteView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Suite",
        "tags": [
          "fingerprints"
        ]
      }
    },
    "/fingerprints/{pair_id}/versions": {
      "get": {
        "operationId": "get_versions_fingerprints__pair_id__versions_get",
        "parameters": [
          {
            "in": "path",
            "name": "pair_id",
            "required": true,
            "schema": {
              "title": "Pair Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FingerprintVersions"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Versions",
        "tags": [
          "fingerprints"
        ]
      }
    },
...
    "/runs": {
      "post": {
        "operationId": "post_run_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Run",
        "tags": [
          "runs"
        ]
      }
    },
...
  }
}